% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.R
\name{cdf}
\alias{cdf}
\title{A discrete random variable x has probability mass function using R}
\usage{
cdf(x)
}
\arguments{
\item{x}{Variable Distribution Sequence}
}
\value{
the probability mass of random variable x
}
\description{
A discrete random variable x has probability mass function using R
}
\examples{
x=c(1,2,2,2,3)/10
cdf(x)
}
