% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic.R
\name{dic}
\alias{dic}
\title{A  Dichotomy Method for Solving Roots of Equations using R}
\usage{
dic(fun, a, b, eps = 1e-05)
}
\arguments{
\item{fun}{the function that needed to solving roots}

\item{a}{endpoint}

\item{b}{endpoint}
}
\value{
the roots of function
}
\description{
A  Dichotomy Method for Solving Roots of Equations using R
}
\examples{
fun<-function(x){x^3-x-1}
dic(fun,3,2)
}
