% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcf.R
\name{MCF}
\alias{MCF}
\title{a function to compute a Monto Carlo estimate of the Beta(3,3) cdf}
\usage{
MCF(a)
}
\arguments{
\item{a}{the endpoint of the intergret}
}
\value{
the solution of the function
}
\description{
a function to compute a Monto Carlo estimate of the Beta(3,3) cdf
}
\examples{
a<-3
MCF(a)
}
